Moved the class back.
