+++ src/main/java/screens/OrderView.java
@@ -3,6 +3,7 @@ package screens;
 import entities.ResidenceType;
 import get_menus_use_case.GetMenusController;
 import order_history_use_case.OrderHistoryInputBoundary;
+import order_use_case.FoodItemDetailsView;
 import order_use_case.OrderController;
 
 import javax.swing.*;
